+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Better color for the tabs text, removed the bold font on selected
 * Add right padding in select to account for the arrow icon space
+* Listing filter selects adapting for the available space
+* Add right padding in select to account for the arrow icon space
 
 ### Fixed