+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+*
+
+### Changed
+
+*
+
+### Fixed
+
+*
+
+## [0.6.0] - 2021-07-30
+
+### Added
+
 * Add new profile component
 * Add `onEndReachedThreshold`, `refreshing`, `loading` and `onFilter` props to listing